hw/nvme: Add direction to NvmeSg
We want to just use iovecs for DMA access in the nvme device in order to
remove the duality that is peppered all over the ctrl.c file. To do this
we need to replace the dma-helper calls that expect QEMUSGList with
IOVec calls that expect QEMUIOVectors.

Do memory mapping for the IOVecs when the memory is of type "DMA" (not
cmb nor pmr) by adding a direction to the NvmeSg struct so we can have
it when mapping memory for iovecs

This is a preparation commit that does not contain the actual change to
iovec memory mapping. But it adds elements that will be used in
subsequent commits.

Signed-off-by: Joel Granados <j.granados@samsung.com>